+++ lib/sheets_v4/api_object_validation.rb
