+++ lib/sheets_v4/api_object_validation/resolve_schema_ref.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module SheetsV4
-  module ValidateApiObjects
+  module ApiObjectValidation
     # Resolve a JSON schema reference to a Google Sheets API schema
     #
     # This class uses the Google Discovery API to get the schemas. Any schema reference
@@ -9,14 +9,14 @@ module SheetsV4
     # name in the Google Discovery API and returning the schema object (as a Hash).
     #
     # This means that `{ "$ref": "cell_data" }` is resolved by returning
-    # `SheetsV4::ValidateApiObjects::LoadSchemas.new(logger:).call['cell_data']`.
+    # `SheetsV4::ApiObjectValidation::LoadSchemas.new(logger:).call['cell_data']`.
     #
-    # An RuntimeError is raised if `SheetsV4::ValidateApiObjects::LoadSchemas.new.call`
+    # An RuntimeError is raised if `SheetsV4::ApiObjectValidation::LoadSchemas.new.call`
     # does not have a key matching the schema name.
     #
     # @example
     #   logger = Logger.new(STDOUT, level: Logger::INFO)
-    #   ref_resolver = SheetsV4::ValidateApiObjects::ResolveSchemaRef.new(logger:)
+    #   ref_resolver = SheetsV4::ApiObjectValidation::ResolveSchemaRef.new(logger:)
     #   people_schema = { 'type' => 'array', 'items' => { '$ref' => 'person' } }
     #   json_validator = JSONSchemer.schema(people_schema, ref_resolver:)
     #   people_json = [{ 'name' => { 'first' => 'John', 'last' => 'Doe' } }]
@@ -62,7 +62,7 @@ module SheetsV4
       def call(ref)
         schema_name = ref.path[1..]
         logger.debug { "Reading schema #{schema_name}" }
-        schemas = SheetsV4::ValidateApiObjects::LoadSchemas.new(logger:).call
+        schemas = SheetsV4::ApiObjectValidation::LoadSchemas.new(logger:).call
         schemas[schema_name].tap do |schema_object|
           raise "Schema for #{ref} not found" unless schema_object
         end
